% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{abandoned}
\alias{abandoned}
\title{Check for Abandoned Tasks}
\usage{
abandoned(cl)
}
\arguments{
\item{cl}{a cluster object}
}
\value{
the number of items on the shelf
}
\description{
Check the shelf for any abandoned tasks
}
