% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_add1_task}
\alias{cluster_add1_task}
\title{Add One Task to a Cluster}
\usage{
cluster_add1_task(cl, task)
}
\arguments{
\item{cl}{a cluster object}

\item{task}{a task object}
}
\description{
Add a task to the input queue of a cluster
}
