% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_create}
\alias{cluster_create}
\title{Create a Cluster}
\usage{
cluster_create(name, mapsize = getOption("threadpool_default_mapsize"))
}
\arguments{
\item{name}{the name of the cluster}

\item{mapsize}{\code{mapsize} argument for underlying LMDB database}
}
\description{
Create the input job queue, the output job queue, and the metadata path
for a cluster
}
