% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_finish_task}
\alias{cluster_finish_task}
\title{Finish a Cluster Task}
\usage{
cluster_finish_task(cl, output)
}
\arguments{
\item{cl}{a cluster object}

\item{output}{the output from a task}
}
\description{
Take the output from running a task and add it to the output queue
}
