% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_initialize}
\alias{cluster_initialize}
\title{Initialize Cluster Input Queue}
\usage{
cluster_initialize(cl_name, x, f, envir = parent.frame())
}
\arguments{
\item{cl_name}{cluster name}

\item{x}{the data}

\item{f}{a function to map to the data}

\item{envir}{an environment within which to evaluate the function \code{f}}
}
\description{
Intialize the cluster input queue by adding all of the jobs based on the
input data
}
