% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_join}
\alias{cluster_join}
\title{Join a Cluster}
\usage{
cluster_join(name)
}
\arguments{
\item{name}{name of the cluster}
}
\value{
A cluster object is returned.
}
\description{
Given a cluster name, join that cluster and return a cluster
object for subsequent passing to \code{cluster_run}.
}
\details{
Join a currently running cluster in order to execute jobs
}
