% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_map}
\alias{cluster_map}
\title{Map a function to data}
\usage{
cluster_map(x, f, cl_name = NULL, ncores = 1L, envir = parent.frame(),
  cleanup = FALSE)
}
\arguments{
\item{x}{the data}

\item{f}{function to be mapped to the data}

\item{cl_name}{cluster name}

\item{ncores}{the number of cores to uses}

\item{envir}{the evaluation environment}

\item{cleanup}{if TRUE, cluster is deleted at the end (default FALSE)}
}
\description{
Cluster version of map to map a function to data elements
}
