% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_next_task}
\alias{cluster_next_task}
\title{Retrieve the Next Task}
\usage{
cluster_next_task(cl)
}
\arguments{
\item{cl}{a cluster object}
}
\value{
a task object
}
\description{
Retrieve the next task in the input queue for a cluster
}
