% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_paths}
\alias{cluster_paths}
\title{Return cluster paths}
\usage{
cluster_paths(name)
}
\description{
The cluster 'name' ends up being a directory on the filesystem. Then, there
separate files for the queues and evaluation environment.
}
