% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_results}
\alias{cluster_results}
\title{Read Results}
\usage{
cluster_results(cl)
}
\arguments{
\item{cl}{cluster object}
}
\value{
a list with the results of the cluster output
}
\description{
Read the results of a cluster run from the output queue
}
