% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_run}
\alias{cluster_run}
\title{Run Tasks in a Cluster}
\usage{
cluster_run(cl, verbose = TRUE)
}
\arguments{
\item{cl}{cluster object}

\item{verbose}{print diagnostic messages?}
}
\value{
the cluster object is returned
}
\description{
This function takes information about a cluster and begins
reading and executing tasks from the associated input queue.
}
\details{
Begin running tasks from a cluster queue
}
