% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_shutdown}
\alias{cluster_shutdown}
\title{Shutdown a Cluster}
\usage{
cluster_shutdown(cl)
}
\arguments{
\item{cl}{a cluster object}
}
\description{
Shutdown a cluster by closing all open threads
}
