% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.R
\name{logfile_show}
\alias{logfile_show}
\title{Show cluster node log file}
\usage{
logfile_show(name, pid)
}
\arguments{
\item{name}{name of cluster}

\item{pid}{process ID for cluster node}
}
\description{
Show the output being sent to the cluster node log file
}
