% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{show_log_file}
\alias{show_log_file}
\title{Show cluster node log file}
\usage{
show_log_file(name, pid)
}
\arguments{
\item{name}{name of cluster}

\item{pid}{process ID for cluster node}
}
\description{
Show the output being sent to the cluster node log file
}
