% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cluster_add_tasks}
\alias{cluster_add_tasks}
\title{Add a Tasks to a Cluster}
\usage{
cluster_add_tasks(cl, x, f)
}
\arguments{
\item{x}{the data}

\item{f}{function to be applied to the data}

\item{cl_name}{a cluster object}
}
\description{
Add a batch of tasks to a cluster
}
