% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_create}
\alias{cluster_create}
\title{Create a Cluster}
\usage{
cluster_create(name)
}
\arguments{
\item{name}{the name of the cluster}
}
\description{
Create the input job queue, the output job queue, and other cluster elements
}
