% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_join}
\alias{cluster_join}
\title{Join a Cluster}
\usage{
cluster_join(name, mapsize = getOption("threadpool_default_mapsize"))
}
\arguments{
\item{name}{name of the cluster}

\item{mapsize}{\code{mapsize} argument for underlying LMDB database}
}
\value{
A cluster object is returned.
}
\description{
Given a cluster name, join that cluster and return a cluster
object for subsequent passing to \code{cluster_run}.
}
\details{
Join a currently running cluster in order to execute jobs
}
