% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{initialize_cluster_queue}
\alias{initialize_cluster_queue}
\title{Initialize Cluster Input Queue}
\usage{
initialize_cluster_queue(cl_name, x, f, meta)
}
\arguments{
\item{cl_name}{cluster name}

\item{x}{the data}

\item{f}{a function to map to the data}

\item{meta}{arbitrary metadata for the applying the function \code{f}}
}
\description{
Intialize the cluster input queue by adding all of the jobs based on the
input data
}
