% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{tp_map}
\alias{tp_map}
\title{Thread Pool Map}
\usage{
tp_map(x, f, cl_name, envir = parent.frame(), ncores = 2L)
}
\arguments{
\item{x}{an R object that will be coerced to a list}

\item{f}{a function to be mapped to the elements of \code{x}}

\item{cl_name}{an optional name for the cluster queue}

\item{envir}{an environment within which to evaluate the function \code{f}}

\item{ncores}{the number of cores to use}
}
\value{
a list containing the results
}
\description{
Map a function in parallal across the elements of a list
}
