% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{tp_map}
\alias{tp_map}
\title{Thread Pool Map}
\usage{
tp_map(x, f, meta = NULL, cl_name = NULL, ncores = 2L,
  wait_for_result = TRUE)
}
\arguments{
\item{x}{an R object that will be coerced to a list}

\item{f}{a function to be mapped to the elements of \code{x}}

\item{meta}{arbitrary metadata for the applying the function \code{f}}

\item{cl_name}{an optional name for the cluster queue}

\item{ncores}{the number of cores to use}

\item{wait_for_result}{should we wait for all results to finish?}
}
\description{
Map a function in parallal across the elements of a list
}
